+++ include/vendors/OceanOptics/protocols/obp/impls/OBPOpticalBenchProtocol.h
@@ -42,8 +42,8 @@ class OBPOpticalBenchProtocol: public OpticalBenchProtocolInterface {
 	OBPOpticalBenchProtocol();
 	virtual ~OBPOpticalBenchProtocol();
 
-	virtual unsigned int readOpticalBenchSlitWidthMicrons(const Bus &bus) throw(ProtocolException);
-	virtual unsigned int readOpticalBenchFiberDiameterMicrons(const Bus &bus) throw(ProtocolException);
+	virtual unsigned short int readOpticalBenchSlitWidthMicrons(const Bus &bus) throw(ProtocolException);
+	virtual unsigned short int readOpticalBenchFiberDiameterMicrons(const Bus &bus) throw(ProtocolException);
 	virtual std::string *readOpticalBenchID(const Bus &bus) throw(ProtocolException);
 	virtual std::string *readOpticalBenchSerialNumber(const Bus &bus) throw(ProtocolException);
 	virtual std::string *readOpticalBenchCoating(const Bus &bus) throw(ProtocolException);
